refactor(ppl-lint): unify edit-distance on Damerau-Levenshtein
The command suggester used Damerau-Levenshtein (transpositions cost 1) while the
field suggester used plain Levenshtein (transpositions cost 2, compensated by a
wider threshold) — two near-identical sweeps that differed only by which typo
class they ranked well. Extract a shared `edit_distance` module exposing
`damerauLevenshtein` and a `nearestWithinThreshold` helper, and route both
`suggestField` and `suggestCommand` through it. Delete the plain `levenshtein`
implementation; both callers now rank transposition typos closer at no extra
cost. `command_suggestion` re-exports `damerauLevenshtein` so existing importers
are unaffected. Behavior is preserved (verified by the existing suggester tests,
including the distance-0 case-preference case).
